Created script to execute the dummy data generation directly
