+++ app/client.go
@@ -142,6 +142,7 @@ func loadV1(c *cli.Context) error {
 	dataFilesPath := c.String("files")
 	sensitiveFilePath := c.String("sensitive")
 	entryPointIdx := c.Int("entryPointIdx")
+	dummy :=  c.Bool("dummy")
 	empty := c.Bool("empty")
 
 	// db settings
@@ -187,6 +188,15 @@ func loadV1(c *cli.Context) error {
 	}
 	directory := filepath.Dir(dataFilesPath)
 
+	//if flag is on -> re-generate dummy patients
+	if dummy {
+		err := loader.ExecuteScript("/scripts/import-tool/generateDummyData.sh " + directory + "/" + files.PatientDimension + " " + directory + "/" + files.ObservationFact + " " + directory + "/" + files.DummyToPatient)
+		if err != nil {
+			log.Error("Error while executing generateDummyData script:", err)
+			return cli.NewExitError(err, 1)
+		}
+	}
+
 	// get the list of sensitiveConcepts
 	f, err = os.Open(sensitiveFilePath)
 	if err != nil {
@@ -207,7 +217,7 @@ func loadV1(c *cli.Context) error {
 		mapSensitive[line] = struct{}{}
 	}
 
-	loaderi2b2.LoadI2B2Data(el.Roster, entryPointIdx, directory, files, allSensitive, mapSensitive, databaseS, empty)
+	err = loaderi2b2.LoadI2B2Data(el.Roster, entryPointIdx, directory, files, allSensitive, mapSensitive, databaseS, empty)
 	if err != nil {
 		log.Error("Error while converting I2B2 data:", err)
 		return cli.NewExitError(err, 1)
