--- loader/commonStructs.go
@@ -1,10 +1,0 @@
-package loader
-
-// DBSettings stores the database settings
-type DBSettings struct {
-	DBhost     string
-	DBport     int
-	DBuser     string
-	DBpassword string
-	DBname     string
-}
